NABTO-1332: shared tunnel functions moved to tunnel_common.c
