+++ src/modules/tunnel/tunnel_select.c
@@ -68,14 +68,14 @@ void tunnel_loop_select() {
 
         for (i = 0; i < NABTO_MEMORY_STREAM_MAX_STREAMS; i++) {
             if (tunnels[i].state != TS_IDLE) {
-                if (tunnels[i].state == TS_FORWARD && tunnels[i].tcpReadState == FS_READ) {
-                    FD_SET(tunnels[i].sock, &read_fds);
-                    max_read_fd = MAX(max_read_fd, tunnels[i].sock);
+                if (tunnels[i].state == TS_FORWARD && tunnels[i].extReadState == FS_READ) {
+                    FD_SET(tunnels[i].tunnel_type_vars.tcp.sock, &read_fds);
+                    max_read_fd = MAX(max_read_fd, tunnels[i].tunnel_type_vars.tcp.sock);
                 }
                 if ((tunnels[i].state == TS_FORWARD && tunnels[i].unabtoReadState == FS_WRITE) ||
                     tunnels[i].state == TS_OPENING_SOCKET) {
-                    FD_SET(tunnels[i].sock, &write_fds);
-                    max_write_fd = MAX(max_write_fd, tunnels[i].sock);
+                    FD_SET(tunnels[i].tunnel_type_vars.tcp.sock, &write_fds);
+                    max_write_fd = MAX(max_write_fd, tunnels[i].tunnel_type_vars.tcp.sock);
                 }
             }
         }
@@ -112,10 +112,10 @@ void tunnel_loop_select() {
 #endif
             unabto_network_select_read_sockets(&read_fds);
             for (i = 0; i < NABTO_MEMORY_STREAM_MAX_STREAMS; i++) {
-                if (tunnels[i].sock != INVALID_SOCKET && FD_ISSET(tunnels[i].sock, &read_fds)) {
+                if (tunnels[i].tunnel_type_vars.tcp.sock != INVALID_SOCKET && FD_ISSET(tunnels[i].tunnel_type_vars.tcp.sock, &read_fds)) {
                     tunnel_event(&tunnels[i], TUNNEL_EVENT_SOURCE_TCP_READ);
                 }
-                if (tunnels[i].sock != INVALID_SOCKET && FD_ISSET(tunnels[i].sock, &write_fds)) {
+                if (tunnels[i].tunnel_type_vars.tcp.sock != INVALID_SOCKET && FD_ISSET(tunnels[i].tunnel_type_vars.tcp.sock, &write_fds)) {
                     tunnel_event(&tunnels[i], TUNNEL_EVENT_SOURCE_TCP_WRITE);
                 }
             }
